% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_data.R
\name{aggregate_data}
\alias{aggregate_data}
\title{convert individual level VB data list to aggregate level}
\usage{
aggregate_data(data)
}
\description{
convert individual level VB data list to aggregate level
}
