% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_data.R
\name{expand_data}
\alias{expand_data}
\title{aggregate_data = function(data) {
  AEs = sort(unique(data$AE_NAME))
  counts <- count(data,Age,Male,VAX_TYPE)
  nn <- counts$n
  A = matrix(0, nrow(counts), length(AEs))
  colnames(A) = AEs
  for(i in 1:length(nn)){
    data_sub = subset(data, Age==counts$Age[i]& 
                        Male==counts$Male[i]& VAX_TYPE==counts$VAX_TYPE[i])
    A[i,] = sapply(1:length(AEs), function(j) sum(data_sub$AE_NAME == AEs[j]))
    
  }
  X = counts %>%
    select(-n) %>%
    as.matrix()
  X = cbind(Intercept = 1, X)
  
  V = X[, 4]
  X = X[, -4]
  return(data=list(A=A, nn=nn, X=X, V=V))
}}
\usage{
expand_data(data)
}
\description{
#' @export
individual_data = function(data) {
  AEs = sort(unique(data$AE_NAME))
  X <- data %>%
          distinct()%>%
          mutate(Intercept = 1) %>%
          select(Intercept, SEX, AGE, VAX_TYPE) %>%
          as.matrix()
  V <- X[, 4]
  X <- X[, -4]
  nn = rep(1, nrow(X))
  
  cores = floor(detectCores() / 1.5)
  cl = makeCluster(cores, setup_strategy = "sequential")
  registerDoParallel(cl)
  A = foreach(i = 1:length(AEs),
              .packages = c("tidyverse"),
              .combine=cbind
              ) %dopar% {
                ifelse(data$AE_NAME == AEs[i], 1, 0)
                }
  stopCluster(cl)
}
\details{

}
