\name{VBVaccine-package}
\alias{VBVaccine-package}
\alias{VBVaccine}
\docType{package}
\title{
\packageTitle{VBVaccine}
}
\description{
\packageDescription{VBVaccine}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{VBVaccine}
\packageIndices{VBVaccine}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{VBVaccine}

Maintainer: \packageMaintainer{VBVaccine}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
